backend/fetchers/web_scraper.py
import requests
from bs4 import BeautifulSoup
from typing import List, Dict, Optional
import time

class WebScraper:
    """Generic web scraper for travel information"""

    def __init__(self):
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36'
        }
        self.session = requests.Session()
        self.session.headers.update(self.headers)

    def scrape_travel_tips(self, destination: str) -> List[str]:
        """Scrape general travel tips (placeholder - can be expanded)"""
        # This is a basic implementation
        # In production, you'd want to use specific APIs or be more careful with scraping
        tips = [
            f"Research {destination} weather patterns before your trip",
            f"Check visa requirements for {destination}",
            f"Learn basic local phrases",
            f"Book accommodations in advance during peak season",
            f"Keep emergency contacts handy"
        ]
        return tips

    def get_page_content(self, url: str) -> Optional[str]:
        """Fetch and parse page content"""
        try:
            response = self.session.get(url, timeout=10)
            response.raise_for_status()

            soup = BeautifulSoup(response.content, 'html.parser')

            # Remove script and style elements
            for script in soup(["script", "style"]):
                script.decompose()

            # Get text
            text = soup.get_text()

            # Clean up text
            lines = (line.strip() for line in text.splitlines())
            chunks = (phrase.strip() for line in lines for phrase in line.split("  "))
            text = ' '.join(chunk for chunk in chunks if chunk)

            return text[:5000]  # Limit to 5000 chars
        except Exception as e:
            print(f"Error scraping {url}: {e}")
            return None

    def search_destination_info(self, destination: str) -> Dict:
        """Search for destination information (basic implementation)"""
        return {
            'destination': destination,
            'tips': self.scrape_travel_tips(destination),
            'note': 'This is a basic scraper. Consider using specific APIs for better data.'
        }
