backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
llama-cpp-python>=0.2.27
requests>=2.31.0
beautifulsoup4>=4.12.3
googlemaps>=4.10.0
wikipedia>=1.4.0
pillow>=10.3.0
markdown>=3.5.2
weasyprint>=61.0
python-multipart>=0.0.6
pydantic>=2.5.3
aiohttp>=3.9.1
lxml>=5.1.0
